Move the UI related classes to the UI package
