+++ app/src/main/java/com/arctouch/codechallenge/ui/HomeAdapter.kt
@@ -1,4 +1,4 @@
-package com.arctouch.codechallenge.home
+package com.arctouch.codechallenge.ui
 
 import android.support.v7.widget.RecyclerView
 import android.view.LayoutInflater
