+++ app/src/main/res/layout/content_details.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     app:layout_behavior="@string/appbar_scrolling_view_behavior"
-    tools:context="com.arctouch.codechallenge.details.DetailsActivity"
+    tools:context="com.arctouch.codechallenge.ui.DetailsActivity"
     tools:showIn="@layout/details_activity">
 
     <android.support.constraint.ConstraintLayout
