+++ app/src/main/res/layout/details_activity.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:fitsSystemWindows="true"
-    tools:context="com.arctouch.codechallenge.details.DetailsActivity">
+    tools:context="com.arctouch.codechallenge.ui.DetailsActivity">
 
     <android.support.design.widget.AppBarLayout
         android:id="@+id/app_bar"
